Move the zip file content checker into Loaders.cpp, should fix the libretro build
