+++ Core/Loaders.cpp
@@ -315,3 +315,166 @@ bool UmdReplace(const Path &filepath, FileLoader **fileLoader, std::string &erro
 	}
 	return true;
 }
+
+bool DetectZipFileContents(const Path &fileName, ZipFileInfo *info) {
+	struct zip *z = ZipOpenPath(fileName);
+	if (!z) {
+		info->contents = ZipFileContents::UNKNOWN;
+		return false;
+	}
+	DetectZipFileContents(z, info);
+	zip_close(z);
+	return true;
+}
+
+static int countSlashes(const std::string &fileName, int *slashLocation) {
+	int slashCount = 0;
+	int lastSlashLocation = -1;
+	if (slashLocation) {
+		*slashLocation = -1;
+	}
+	for (size_t i = 0; i < fileName.size(); i++) {
+		if (fileName[i] == '/') {
+			slashCount++;
+			if (slashLocation) {
+				*slashLocation = lastSlashLocation;
+				lastSlashLocation = (int)i;
+			}
+		}
+	}
+
+	return slashCount;
+}
+
+inline char asciitolower(char in) {
+	if (in <= 'Z' && in >= 'A')
+		return in - ('Z' - 'z');
+	return in;
+}
+
+void DetectZipFileContents(struct zip *z, ZipFileInfo *info) {
+	int numFiles = zip_get_num_files(z);
+	_dbg_assert_(numFiles >= 0);
+
+	// Verify that this is a PSP zip file with the correct layout. We also try
+	// to detect simple zipped ISO files, those we'll just "install" to the current
+	// directory of the Games tab (where else?).
+	bool isPSPMemstickGame = false;
+	bool isZippedISO = false;
+	bool isTexturePack = false;
+	bool isFrameDump = false;
+	int stripChars = 0;
+	int isoFileIndex = -1;
+	int stripCharsTexturePack = -1;
+	int textureIniIndex = -1;
+	int filesInRoot = 0;
+	int directoriesInRoot = 0;
+	bool hasParamSFO = false;
+	bool hasIcon0PNG = false;
+	s64 totalFileSize = 0;
+
+	// TODO: It might be cleaner to write separate detection functions, but this big loop doing it all at once
+	// is quite convenient and makes it easy to add shared heuristics.
+	for (int i = 0; i < numFiles; i++) {
+		const char *fn = zip_get_name(z, i, 0);
+
+		zip_stat_t stat{};
+		zip_stat_index(z, i, 0, &stat);
+		totalFileSize += stat.size;
+
+		std::string zippedName = fn;
+		std::transform(zippedName.begin(), zippedName.end(), zippedName.begin(),
+			[](unsigned char c) { return asciitolower(c); });  // Not using std::tolower to avoid Turkish I->ı conversion.
+		// Ignore macos metadata stuff
+		if (startsWith(zippedName, "__macosx/")) {
+			continue;
+		}
+		if (endsWith(zippedName, "/")) {
+			// A directory. Not all zips bother including these.
+			continue;
+		}
+
+		int prevSlashLocation = -1;
+		int slashCount = countSlashes(zippedName, &prevSlashLocation);
+		if (zippedName.find("eboot.pbp") != std::string::npos) {
+			if (slashCount >= 1 && (!isPSPMemstickGame || prevSlashLocation < stripChars + 1)) {
+				stripChars = prevSlashLocation + 1;
+				isPSPMemstickGame = true;
+			} else {
+				INFO_LOG(Log::HLE, "Wrong number of slashes (%i) in '%s'", slashCount, fn);
+			}
+			// TODO: Extract icon and param.sfo from the pbp to be able to display it on the install screen.
+		} else if (endsWith(zippedName, ".iso") || endsWith(zippedName, ".cso") || endsWith(zippedName, ".chd")) {
+			if (slashCount <= 1) {
+				// We only do this if the ISO file is in the root or one level down.
+				isZippedISO = true;
+				INFO_LOG(Log::HLE, "ISO found in zip: %s", zippedName.c_str());
+				if (isoFileIndex != -1) {
+					INFO_LOG(Log::HLE, "More than one ISO file found in zip. Ignoring additional ones.");
+				} else {
+					isoFileIndex = i;
+					info->contentName = zippedName;
+				}
+			}
+		} else if (zippedName.find("textures.ini") != std::string::npos) {
+			int slashLocation = (int)zippedName.find_last_of('/');
+			if (stripCharsTexturePack == -1 || slashLocation < stripCharsTexturePack + 1) {
+				stripCharsTexturePack = slashLocation + 1;
+				isTexturePack = true;
+				textureIniIndex = i;
+			}
+		} else if (endsWith(zippedName, ".ppdmp")) {
+			isFrameDump = true;
+			isoFileIndex = i;
+			info->contentName = zippedName;
+		} else if (endsWith(zippedName, "/param.sfo")) {
+			// Get the game name so we can display it.
+			std::string paramSFOContents;
+			if (ZipExtractFileToMemory(z, i, &paramSFOContents)) {
+				ParamSFOData sfo;
+				if (sfo.ReadSFO((const u8 *)paramSFOContents.data(), paramSFOContents.size())) {
+					if (sfo.HasKey("TITLE")) {
+						info->gameTitle = sfo.GetValueString("TITLE");
+						info->savedataTitle = sfo.GetValueString("SAVEDATA_TITLE");
+						char buff[20];
+						strftime(buff, 20, "%Y-%m-%d %H:%M:%S", localtime(&stat.mtime));
+						info->mTime = buff;
+						info->savedataDetails = sfo.GetValueString("SAVEDATA_DETAIL");
+						info->savedataDir = sfo.GetValueString("SAVEDATA_DIRECTORY");  // should also be parsable from the path.
+						hasParamSFO = true;
+					}
+				}
+			}
+		} else if (endsWith(zippedName, "/icon0.png")) {
+			hasIcon0PNG = true;
+		}
+		if (slashCount == 0) {
+			filesInRoot++;
+		}
+	}
+
+	info->stripChars = stripChars;
+	info->numFiles = numFiles;
+	info->isoFileIndex = isoFileIndex;
+	info->textureIniIndex = textureIniIndex;
+	info->ignoreMetaFiles = false;
+	info->totalFileSize = totalFileSize;
+
+	// Priority ordering for detecting the various kinds of zip file content.s
+	if (isPSPMemstickGame) {
+		info->contents = ZipFileContents::PSP_GAME_DIR;
+	} else if (isZippedISO) {
+		info->contents = ZipFileContents::ISO_FILE;
+	} else if (isTexturePack) {
+		info->stripChars = stripCharsTexturePack;
+		info->ignoreMetaFiles = true;
+		info->contents = ZipFileContents::TEXTURE_PACK;
+	} else if (stripChars == 0 && filesInRoot == 0 && hasParamSFO && hasIcon0PNG) {
+		// As downloaded from GameFAQs, for example.
+		info->contents = ZipFileContents::SAVE_DATA;
+	} else if (isFrameDump) {
+		info->contents = ZipFileContents::FRAME_DUMP;
+	} else {
+		info->contents = ZipFileContents::UNKNOWN;
+	}
+}
